Move split resource definitions into new package
